Refactor delete_owner_password_rules/1 rename, rtn
This is smaller refactoring of the previously named
delete_password_rules/1 function.  This function was missed when we made
the owner related calls include `owner` in the function name to be
consistent with the network rules.  So this rename is one change here.

The other change is in the return values from this function (and related
typespec).  Previously we simply returned `:ok` regardless of the action
actually taken (assuming no exception).  This is relatively information
free and not consistent with some recent places where we're explicit
about whether or not we actually deleted anything.  So this change makes
this call consistent with that more informative standard.