Refactor create_identity callback return & similar
This is similar to the result tuple return additions done with things
like credentials and rules.  Turns out that we may well be calling some
of these functions in the ExtendedAuthLogic and ExtendedMgmtLogic
modules using `with` expressions and the result tuple is just a bit more
sensible than not having a result tuple in these cases.

Update across Identity types and for types which resemble or include
calls to create_identity/3.