Move ExtendedAuthLogic static defaults to modattrs
Static/literals such as option defaults should be captured as module
attributes towards the top of the file so that: 1) they are defined in
one place, and 2) that any necessary changes to the values are easy to
maintain since hunting through the code won't be necessary.

This will be done more broadly and explained again.