Add test for operation clearing in password authn
Email/Password authentication is an interruptable/resumable process in
cases where intermediate user interactions are required: this happens
when an authentication Instance isn't initially defined or MFA is
required.

This commit allows the authenticate_email_password/2 to handle the
resumption of the authentication process by simply supplying the missing
information as part of the resumption options.  This way we don't need
an alternative function or manual intervention in the
authentication_status data to effect a resumption.  This current
addition is only handling Instance identification since MFA is not yet
implemented and there are no other interruptions defined.