Refactor ExtendedAuthLogic Identity handling logic
Prior to this update, Identity processing functions assumed that for an
Identity to be valid, the SystIdentities.validated field had to be non-
null.  This is correct for Email Identities which need a user to
successfully execute the validation confirmation process prior to using
the Email Identity, but other Identity types don't require such
validation.  The existing functionality therefore couldn't successfully
process non-Email Identities.

Earlier a decision was taken that the SystIdentities.validated field
would strictly represent the timestamp when an Identity was explicitly
validated.  This means that implicitly valid Identities would be have a
null validated value since they are never explicitly validated.  Rather
than make the meaning of the validated field context dependent, we'll
change the application logic to allow an identity to be presumed
validated via a function call.