Refactor error tuple formatting
Somewhere along the way a non-standard way of formatting the error
tuples got into the code base and the formatter cannot correct it
(naturally, this is due to cut & pasting).

This commit is not a concerted effort at resolving this issue.  I'm
simply correcting it as its noticed.

In addition other minor fixes to formatting or messaging.