Refactor AccountCode reset & support; add lookup
This commit changes the way Account Code resetting works to be less
strict about their having to be an existing Account Code prior to
attempting.  Since an Account Code is a non-authenticating alternative
way to identity an access account, we can afford to be a little more
accommodating to the caller request.  So we make the functions here a
little less stringent and change the supporting Impl.Identity
functionality to more cleanly handle cases where we might not have an
existing identity to delete.

  - Allow the reset function to either do a full delete/create cycle if
    a true reset is being requested or just a create cycle if a prior
    Account Code doesn't exist.

  - Change the Impl.Identity.delete_identity/1 function to not query so
    much.  This was the old retrieve the record so we could pass it to
    delete/2.  Now we just delete_all/2 on the ID.

  - Change Impl.Identity.delete_identity/1 to return :not_found if an
    attempted delete doesn't actually do anything rather than raising.
    Truth is most of the time such a delete won't be a fatal exception
    and the caller really should make that choice.

  - Add AccountCode.get_account_code_by_access_account_id/1 to support
    retrieving the Account Code Identity record by Access Account ID.