Refactor Impl.Identity.delete_identity/1
This update implements two changes: 1) change the return values to be
:deleted | :not_found rather than :ok | :not_found; 2) require that the
identity type be named making the function delete_identity/2.

The return type change is simply to be more consistent with how the
usage has evolved relative to other deleting functionality in the
component.

The reason to require the identity type at delete time is to ensure that
the correct identity type specific delete/revoke function is called to
process the request; prior to this commit it was possible to call the
ApiToken version of delete identity to delete an Email Identity for
example.  This is be bad because specialized logic could be missed or
authorization for one type could be coerced into a more general
capability.  There are a couple other instances where this should be
addressed (not just deletes), but this is the more urgent of the
instances to remedy.