Refactor email validation timing in Authentication
Prior to this commit.  The email as submitted was passed to the
Identifier Rate Limit checks and then later submitted for normalization
just prior to the Identity record lookup.  The problem with this is that
it could allow an attacker to bypass rate limits for an identifier by
submitting different de-normalized forms of the same identifier which
would be normalized after the rate limit check; the severity of the
issue would vary based on the size of the domain part of an email
address which is case insensitive.

This commit includes the needed changes to allow an Email address
identifier to be normalized prior to checking the identifier rate limit
which means any equivalent email identifier, regardless of form
submitted, will be tested against a single rate limit counter.

There is a known issue in that we'll double validate and normalize the
address, first prior to rate limit check and next where we were doing
those steps originally in the identification process post-rate limit
check... but that's a minor concern right now and doesn't waste that
many cycles.  A good to-do for later.