Fix incorrect rate limiter counter_id construction
We were constructing the counter_id value prior to calling the rate
limit functions... which themselves constructing the counter_id.  This
meant that the counter type was being represented twice in the counter
name rather than once.

Remove the unnecessary instances of constructing the counter_id prior to
rate limiting function calls.