Fix Credo reraise warnings & error tuple formats
When raising from within a `rescue` block, Credo warns that we should be
reraising to preserve the stacktrace.  I'll buy it and I don't see that
it hurts beyond any continuing ills around our pervasive rescue calls
and possibly adding to our credential logging potential... again an
existing/known issue.

At some point we really need to revisit our approach to how we deal with
errors.  Do we do as we do and rescue/log/reraise/error tuple as much as
we are right now?  Do we just let most things blow up or only try to
push these captures even closer to the API definition?  Not a problem
for now, but something to think about.