Fix Authentication "never match" Dialyzer report
Actually, I still get that error from the ElixirLS Dialyzer, but not my
primary Dialyzer run.  Not sure why it's this is still complaining,
especially complaining in one Dialyzer context but not another, but for
now it's good enough.

(Yes, the raise in the error branch probably meets the definition of
not returning... but it shouldn't: I don't want a return there).