Refactor Identity validated field use
Previously, I had decided that the syst_identities.validated field would
be used only to indicate that explicitly "validated" Identities, such as
Email Identities, had in fact been validated; this means that implicitly
valid Identities would keep this field NULL and simply not test for
validity.

After thinking it over a bit, that decision was just wrong.  There is
some value in having a field indicate that an Identity is explicitly
validated and when a real user decided to confirm that validation, but
the benefits of that don't make up for the costs of making that happen.
By making the validated field mean valid for use generally, and not just
Email Identities as the previous decision had it, we have to know less
lore/special cases about the field usage and reduce that to if the field
is null you cannot use the Identity and if its populated it's valid for
use (assuming there's no other invalidator).  This makes the reporting
of explicit setting vs. implicit setting the special case which only
matters in reporting, not in functionality.  A much safer alternative.

So make it so that all identities must have the validated field set to
be usable and for those that are implicitly validated, set their
create_validated option default to be true.

Also, a few places were found were we weren't defining option defaults
via module attributes; make that change where found.

Update test data and any impacted tests to reflected the changed
validated field decision.