Refactor ExtendedMgmtLogic funcs into DB Trans.
Some functions in the ExtendedMgmtLogic module combined data changing
logic into single function calls, but didn't wrap those calls into a
database transaction when they really ought to have.  This commit fixes
that.

Some functionality wrapped multi-step data changes farther down the call
stack and in those cases no special handling was made in this module.
The thinking is to keep the transactions close to the code that
necessitates a database transaction.  This isn't a perfect solution but
one I believe addresses the most pressing needs; specifically, removing
the transaction management to a distant function means we could bypass
the transaction handling by choosing a different entry point to the
underlying logic.  The primary downside to the favored approach is that
transaction handling may appear at different levels of abstraction which
could make it easy to miss when required.  There is a secondary downside
in that we may more commonly nest transactions when we don't have to,
but that won't be harmful in most cases, adding only some minor
processing overhead.

Another issue raised by the changes here is really how we want to
abstract error conditions.  The current form of caught errors is done
with the idea that we will capture a stack trace of sorts from the root
cause through to higher levels to the point of call.  This worked well
in the JavaScript world that we previously wrote extensively for but
this is a different environment and different rules may apply.  The
code committed here breaks with some of that thinking to destructure
error conditions, discarding intermediate error messages and directly
presenting the root cause as the immediate cause.  This certainly makes
sense here since a root cause for Recovery requests can be that a
recovery is already underway, something we want a caller to be able to
figure out without much ceremony.  Anyway, this is a larger issue to
think about and we're getting close to where we can make a more sensible
decision on the issue.