Fix MsbmsSystAuthentication Credo AliasUsage
Per CI.  The Credo.Check.Design.AliasUsage we were using was checking
for a depth of 2.  This was causing a whole slew of Credo warnings which
causes CI to fail.

Set `if_nested_deeper_than` to 3.  The CI failure was actually expected
and the current decision in that our aliasing should be very shallow,
really just getting us out of specifying the Component name.   This is
because we want the source path of the module to be apparent in the
usage and aliasing deep into module name can obscure that.

Not completely sure that this is the right decision, but we'll go with
it for now so we can make progress on more important things.