Rename global/instance to msmcp/msbms
This is a more general renaming related to what is now the Master
Control Program (global) and the MSBMS (instance).  So in addition to
changing the name of these two larger abstractions, we also true up the
names committed earlier since much of the search/replace refactoring
skipped this top level Elixir app and some database definition files.