Rename non-"msbms" specific names to be "ms"
This is mostly about database schema objects and the related scripts/
directory names.  Since most of the objects developed up to this point
are genuinely generic in applicability (not necessarily tied to the idea
of MSBMS), many things named as being "MSBMS" really should be more
generic.

This is especially true of database namespaces since there we'll end up
mixing general relations (e.g. ms_syst_data.syst_enum_items) with the
more specific.  This could become a hassle if we have both ms_syst_data
and msbms_syst_data; so for database organization stick to the more
generic names since the database definition itself will be limiting in
all the important ways to know what's what.

A few fixes and catches along the way so this includes some more general
clean up as well.