Fix MscmpSystInstance.Runtime.Application param
The stop_instance/2 function was getting an options keyword list param,
but then passing it to MscmpSystDb.stop_datastore/2 which was expecting
an optional timeout... but not a full options keyword list.  This was
causing Dialyzer to assume that the options list itself must be a
non_neg_integer and then when it was properly passed to other later
functions... causing Dialyzer to say those calls were in error due to
the type of the options value.  Sigh.  Something was wrong, but not what
Dialyzer was reporting.

Fix the root cause of the issue and Dialyzer is now happy.