Add MCP password and pool size to startup options
This adds a way to specify a password fragment for use in creating
passwords for the MCP Datastore Contexts which access the MCP database.

Also we add a way to designate the number of database connections to
open to the MCP database.

Side note.  All of this Datastore specific stuff does not belong in
MscmpSystOptions; it all is just helper stuff that belongs in
MscmpSystDb... or somewhere else, maybe even it's own component if we
want to keep MscmpSystOptions and MscmpSystDb unaware of each other.

I think we'll do that as part of this feature branch.