Rename set_datastore_context/1 function
Rename MscmpSystDb.set_datastore_context/1 to put_datastore_context/1.

This is done for consistency.  The underlying call is put_dynamic_repo
and we're using put_* for things like setting the enums and settings
names in the process dictionary.  In light of this whole ecosystem
consistency the original rationale of using a different naming
convention to put some distance between MscmpSystDb and Ecto doesn't
really make sense.

Update all calls and documentation as well.