Refactor all Data structs to Msdata namespace
Prior to this point all database references were made via structs in
the <Component Name>.Data namespace.  This wasn't ideal as the structs
essentially introduced secondary public API like entry points even
though we were trying to keep all public API calls in the module named
after the component in question.  When we get to subsystems, however,
this is even worse, because we want the subsystem to abstract away/
decouple the Component layer from the high level application
functionality.  This leaves us with less than satisfactory choices such
as just forcing the higher level application to be aware of all the
components (coupling) or put in an artificial translation in the
subsystem layer.  Neither the extra code or the tighter coupling of high
level implementation with low level implementation seems appealing.

The solution I've landed on for now is to move all of the data structs
into a generically named root namespace, Msdata, that get shared across
components for the data layer.  This isn't a terrible solution.  For now
we can continue to treat the data structs as a special instance of
public API, we can keep the data structs of each component in the source
tree for that component, we get the advantage that data structs must
have unique names (that wasn't true before this change, though
intended), and if we choose to refactor the data out of the components
we can always use the root namespace as the basis of the new data
oriented component (this of course assumes a certain kind of refactoring
as being the most likely refactoring).

The downsides are that we lose the intuitive indication of where a given
data struct actually originates and we break our naming conventions by
having a root module name (Msdata) that isn't descriptive of the
component where the struct is defined.  We'll ignore that I'm electing
to break our three character indicator in the name (msXXX is the
standard and this is msYYYY).  But data is more intuitive than "msdat"
and we're almost always shortening names even with the four character
kind indicator in the name.  Seems worth it.