Refactor MscmpSystAuthn for Enum retrieval funcs
Similar to the way we've added functions to front enumerations for the
MscmpSystInstance component, we do the same here for MscmpSystAuthn.

The basic idea is that we want to hide some of the inside baseball of
enumeration retrieval from the subsystem level and so have dedicated
functions which in turn call the more general MscmpSystEnums functions
for retrieving Enum Items by name or default Enum Item records.