Add MssubMcp processing module
MCP operation must execute in processes where the Enum, Settings, and
MCP Datastore services have been set in the process dictionary.  Much
of the time, since the plan right now is for MCP to be managed by users
in a BMS instance, that context won't be appropriate for just setting
and forgetting.

This module adds functions for dealing with setting the process
dictionary values appropriate to the running MCP settings.  In
particular we add a function which takes a function to execute in that
process context and then reset the process dictionary values back to
whatever they were prior to the operation.

There's some naiveté around this implementation, but it will work for
now.