Add MssubMcp testing support
Create basic test infrastructure for MssubMcp unit and integration
testing.

This update includes a change in MscmpSystInstance unit test seed data
required to avoid naming collisions with production seed data in
MssubMcp.

Mix also introduces elixirc_path support for DevSupport, but that hasn't
been fully implemented yet so as of this commit it's a no-op.