Fix Instance stop_all_applications/1 test failure
In the MssubMcp integration tests, and only in that test, a call to
MscmpSystInstance.stop_all_applications/1 was resulting in an unexpected
shutdown of one of the Supervisors.  At this point the root cause is
still unknown.

Adding a little extra processing/time however seems to make a
difference for some unknown reason.  Given the low criticality of the
issue and the difficulty in tracing it to a root cause, we're going to
hack over it for now.

Add that little bit of extra processing/time and let the test pass for
now.  Also add a note so we don't forget.

Finally, an appeal for ideas was requested at:
https://elixirforum.com/t/dynamicsupervisor-stop-issue/52091