Refactor MscmpSystAuthn integration test enum call
The integration test should use any defined public API where it exists
rather than dependency APIs.  In this case we recently defined a
MscmpSystAuthn API call for default Access Account State, but this part
of the integration test still called the MscmpSystEnums API to get the
value.

Change that to the newer MscmpSystAuthn call.