Refactor mnesia create_schema handling
This was work done to deal with the mnesia start-up issue the rate
limiter was having.

The work in this commit didn't help that issue in the end (it was a
configuration typing issue, but the error message was very unclear).
However what was done here would likely have popped up anyway since the
move to the `with` expression in `setup_rate_limiter/0`.  That process
could have seen an already existing mnesia schema and complained about
it.

So preempt that possibility and keep the more robust code needed or not.