Refactor Elixir project documentation generation
This change is mostly related to moving the contents of Public API
moduledocs to the README.md files of each project.  This way we can see
the top level documentation in places like GitHub or in other tools
expecting to see a README.md file at the top of a project

Some very minor updates to those README.md docs were made, mostly
related to improving the opening text of components which are currently
under-documented.