Add authentication state deadlines
The Email/Password authentication process can be interrupted to collect
required data from user interaction, such as the user selecting an
Instance if the Access Account is an unowned, global account or for the
user to provide MFA credentials.  While we need to wait for user input,
we shouldn't wait forever.

This commit ensures that a resumed authentication process has been
resumed in a reasonable time.  If not, the authentication state is
considered expired and a code indicating that the authentication has
been rejected due to an expired deadline is returned to the caller as
the final authentication state status value.

While this is really only relevant for Email/Password authentication,
the deadline check and options have been added to all authentication
types just for thoroughness and to suggest that it may be required
should a similar interruptable process be implemented.

Update the code, Public API documentation, integration tests, and
regenerate the module documentation.

Also included in this commit, in ExtendedAuthLogic, is a conversion from
direct enumeration look-ups for identity_types to the new internal API
which covers this functionality.