Refactor MssubBms database file organization
This was originally created very early in the project and the structure
from that time just didn't go to where we're at now.

This restructuring is based on the idea that the Elixir components will
line up with the way we've organized the database files.  It's still
pretty early days yet so it may all yet change.  For now, the assumption
is that we'll be doing things pretty granularly in terms of what makes
for a distinct component.  We might be too granular.  The alternative is
at the application module level, so mscmp_brm for example.  There seems
to be no sensible middle group because once you get some dependencies
you quickly just need them all to have a sensible system.  At the very
granular level there are still components which you could legitimately
use independently (not that that's likely).

Anyway, its a starting point.