Add baseline MssubBms Elixir project
While not exactly a clean skeleton start, it's only that plus the
typical first step things to do on a new project.

This does also include a first cut MssubBms database migration.  It's
clear that this is a long way off from what the final will be, but we
need something to start testing with and establishing a decent baseline
of "what is the app" DB schema may as well start here.