Refactor MssubMcp API & Runtime Assumptions
Much to unpack in this commit.

The first thing to address is we preliminarily settle a debate about
where the primary runtime boundaries lie in regard to what is just a
collection of supervisors and services and where we should be defining
OTP Applications.  In this case the subsystem level becomes the level of
the OTP Application.  This means we make firmly establish what processes
exist, how they're identified or accessed, how they are started and
updated, and how everything is bootstrapped.  The original think was to
leave many of these decisions to the top level application for the
maximum flexibility, but in retrospect that's not really necessary and
adds additional complexity.  By the time we have a subsystem defined we
can have stronger opinions about what the runtime properties should be
that the subsystem supports, leaving the top level application to really
be a consumer of the services provided as is.  My feeling right now is
that this will leave the supervisory trees and service access in the
correct state that we'll want to see in the overall running application
into which the subsystem has been incorporated.  I feel reasonably
confident that this is a correct level of abstraction, though working
problems of a more distributed architecture and runtime environment
could still change my mind on this.

The second item addressed is that MssubMcp now wraps the APIs of its
components and offers its own Public API for those component functions.
This I am much less confident about; indeed my gut feeling is that this
is ultimately going to be proved an incorrect approach, but there are
some positive arguments that do support the approach that are currently
driving the decision.  It was looking like there could be a fair amount
of ceremony in executing component processes in the MCP Subsystem
runtime service context (correct database, correct enums service,
correct settings service).  By wrapping the component APIs in calls
which ensure the correct service context is set and then returned to the
prior state, we can remove that burden from the calling code.  We also
relieve the development process of the top level application from the
cognitive load of understanding all of the underlying components and
their individual APIs. And then I would also expect that the subsystem
API could remain more stable over time where component APIs may need to
be more volatile.  On the downside we have increased indirection between
the top level caller and the code actually doing the work; we have code
that only serves as wrapper code to maintain or add when a component
does change in a way that needs to be surfaced at the subsystem API
level; probably small performance impacts to manage the indirection;
leaks in the abstraction, such as not eliding the component types &
typespecs which would cause even more duplication or extra maintenance
to address.  The work is done and I'm going to see if there is some
value in it or not by working the application proper.  At the time the
decision was taken some of the service context support from the
subsystem wasn't there and was only developed to make the wrapped API
easier... but I think those components may make the feared ceremony of
the top level caller dealing with the service context moot and the
component API wrapping unnecessary; this includes the mcp_opfn macro
(terrible name, will probably change), the start_mcp_service_context &
stop_mcp_service_context functions, and the process_operation function.

Anyway this commit makes these changes and also sets up the
documentation required for the wrapped API.  Tests have been updated as
well, though we do not focus on testing the wrapped API beyond the
incidental integration test usages.  This is something to be rectified
should the commitment to the wrapped API become stronger (I'm kinda
hoping I end up ripping it out).