Rename Msbms to MsappPlatform
This should remove ambiguity about having a top level application called
BMS and a subsystem which actually implements the BMS.  The truth is
that (ideally) the top level app will process the UI for BMS and any
other app subsystem that we might add.  The reality is that might not
happen the way we want it for release 1, but it's at least a goal.

This is more of a recreate than a rename, but given that nothing was
really done outside of configuration activities (now duplicated in
MsappPlatform) the distinction due to process probably doesn't matter.