Refactor MscmpSystInstance Application maintenance
Prior to this commit an assumption was made that Elixir code wouldn't be
permitted to maintain ms_syst_data.syst_applications (and related) data
and that this maintenance would occur via database migrations.  This
point of view was mistaken however unless our abstraction layers were to
be violated.  An Application implementing subsystem cannot migrate data
to the MCP subsystem directly without invoking fairly low level database
related APIs to get a context connection with the correct privileges in
the Elixir code.  We could create a standard migration in the MCP
subsystem, but that breaks our abstraction in a different way since now
MCP subsystem would have to be aware of Application subsystems in its
code.

The real solution is to drop our assumption that Elixir code won't
maintain Application & Application Context records and provide an API
for that.  This API is not intended to facilitate direct end user
maintenance of the data, but to allow the Application implementing
subsystems to maintain their own representation in the MCP subsystem.
This should keep the abstractions intact at the price of a small module
implementing a behaviour that MsappPlatform can rely upon to treat all
Application subsystems the same.

This commit implements the Application & Application Context related
maintenance functionality, the required APIs, tests, and documentation.
Probably a little bit of cleanup along the way as well.