Add component MscmpSystAppSubsystem
This is a very small utility component currently targeted at defining a
behaviour for Application Subsystem projects to implement.  This should
allow a common set of services that MsappPlatform can count on for
managing its child application subsystems.

This is just barely more than an Elixir project skeleton with a few
common starting configurations implemented.