Rename database seed data files remove noise words
With everything named "initialized_*" nothing really needed to be so we
remove that here.  The original thinking was that right now it's true we
are initializing everything, but at some point we might have both
initializing seed data and then alterations to data (or non-
initializing seed-style data).  So while just extra keystrokes now, that
filename word could be distinguishing later.  Still, we can make the
assumed case be initialization and then later updates can be
specifically called out (or we can go through this renaming exercise
again.)  For now, let's go with less typing and bet that some fair
portion of this will never need the distinguishing naming anyway.