Remove MscmpSystLimiter explicit initialization
Prior to this commit the component required the calling code to ensure
that the init_rate_limiter/1 function was called to create the Mnesia
table if required.  This was a direct reflection of the Hammer
implementation which required the same thing.

This led to a race condition between the starting of the limiter service
(which set application configuration if not explicitly set) and the init
function (which needed the configuration) call since the limiter service
startup was asynchronous.

Thinking about it, why force the manual creation of the Mnesia table?
If you're going to use MscmpSystLimiter, you're going to always take
that init step anyway; if we could run it as part of the service startup
(and synchronously within that startup) we'd avoid the race condition,
we'd have a cleaner API for the Limiter component, and we wouldn't
actually lose anything.  We were already initializing on an "if not
exists" basis anyway.  The one configuration point that the init
function took can also be passed along with the startup options (and
would have to be anyway even if we wanted a separate init).

So this commit updates MscmpSystLimiter and MscmpSystAuthn (a consumer).
MssubMcp will also have some changes, but those will be included with a
much larger restructuring already done there.