Refactor MssubMcp away from OTP Application
Prior to this commit, MssubMcp was built to be a self-starting OTP
Application proper, including bootstrapping the database and starting
all offered services.  While this was not the early, original design, it
was nonetheless thought that maybe things could be simplified for higher
level code to make the subsystem a runtime unit that could simply be
there to provided services without much fuss.  Turned out this wasn't
the case.

At the MsappPlatform level, where different subsystems and components
are brought together into the culmination of the system as a whole,
conflicts arose where the boundaries between different subsystems and
components either needed to be blurred and API surfaces increased, or a
rethink of the subsystem level as being an independent OTP Application
needed to be considered.

The first issue to deal with was that MssubMcp would eventually need
knowledge about other Application Subsystems if it were to stay an
independent OTP Application; there was no good way to tell MssubMcp
about the other Application subsystem(s).  It felt wrong to include
those in MssubMcp directly, and so API was built (SystApplication and
related) to allow MssubMcp to be updated after the fact.  The next issue
was that the MsappPlatform was itself going to use the MCP as the data
store for configurations and settings, but MsappPlatform had no way to
inject these into the MssubMcp data.  So this would have meant more API
with messy data bootstrapping logic or again breaking separation of
concerns (API was not built).  Finally, Application Subsystems such as
BMS would need to provide information to MsappPlatform so MsappPlatform
could then use logic to bootstrap that into MssubMcp.  This was all too
just much.

This change takes us back to the day 1 design where the top level would
be where all the components come together to form a whole means the
abstraction boundaries between individual components aren't as
troublesome.  To start, MsappPlatform now defines as hosts the data
migrations that MssubMcp and the Application Subsystems (e.g. BMS); the
components and subsystems still define their subplans including always
added seed data, but MsappPlatform can now also define seed data,
privileges, and the like.  Being able to define all the database
requirements at this highest level resolves all of the boundary crossing
data issues that was going to be cumbersome to deal with using extended
APIs and application logic to deal with.  Also, having the subsystem
data migrations defined in different places was going to be more
confusing.  Having the applicable migrations side-by-side at the top
level should make it clearer what's install-able into the database.  The
service startup process and orchestration also is clearer since we're
explicitly starting component and subsystem services and can more easily
ensure that dependencies between services are respected.