Remove underspecs from MssubMcp Dialyzer warnings
Dialyzer was giving a number of warnings related to the new service
startup recently implemented.  They were all of the form that the spec
is a supertype of:

-------------
contract_supertype
Type specification is a supertype of the success typing.
-------------

Which is described by:

-------------
The @spec, while not incorrect, is more general than the type
returned by the function.

    defmodule Example do
      @spec ok() :: any
      def ok() do
        :ok
      end
    end
-------------

Dialyzer is not wrong about this, but its not being particularly helpful
either.  Yes, the defined specs... currently... are more general, but
realistic implementation changes could expand what the code can actually
produce.  So in this case, were setting an expectation for client
callers that they should be ready to handle these possible outcomes even
if they aren't possible (or unlikely as the case is more commonly to be)
today.

For now then, unless there's a compelling reason to be stricter, just
turn off the underspecs warning.