Add component MscmpSystPerms
This component will serve as the means by which to define and manage
available system and application permissions and their grouping into
permission roles.  This component will not deal with authorization
directly; authorization components will make use of this component for
permission definition while defining their own assignment of permission
roles to users.

This commit is a skeleton commit with most of the boiler plate put in
place.

Note that there are a few places where MscmpSystEnums related code has
been left in the files, but commented out.  I've not yet got a use for
enumerations here, but there's at least even odds that it will show up
at some point so don't outright remove those bits just yet.